Created a Seperate file For My Search Function, Which is the Function I want Mr Ryan Justin to mark
