+++ public/index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
     <script src="js/get-book.js"></script>
+    <script src="js/search-books.js"></script>
 
 
 <head>
